add running requirements clion/cmake+make without cuda
