+++ .gitignore
@@ -37,3 +37,4 @@ release
 release/armadillo.obj release/CMakeCache.txt release/CMakeFiles release/cmake_install.cmake release/_deps release/Makefile release/PD release/temp.1.edge release/temp.1.ele release/temp.1.face release/temp.1.node release/temp.poly
 cmake-build-debug/build.ninja cmake-build-debug/CMakeCache.txt cmake-build-debug/CMakeFiles cmake-build-debug/cmake_install.cmake cmake-build-debug/_deps cmake-build-debug/Testing
 temp/CUDAMatrixVectorMult.cpp temp/CUDAMatrixVectorMult.h temp/doubleToFloatDeviceCpy.cuh
+cmake-build*
