--- include/doubleToFloatDeviceCpy.cuh
@@ -1,2 +1,0 @@
-void doubleToFloatDeviceCpy(int n, int coord, double* source, float* target);
-void elementWiseMultiply(int n, double* a, double* b);
