+++ temp/doubleToFloatDeviceCpy.cuh
@@ -1,0 +1,2 @@
+void doubleToFloatDeviceCpy(int n, int coord, double* source, float* target);
+void elementWiseMultiply(int n, double* a, double* b);
